+++ web/js/antragsgruen-en.js
@@ -38,6 +38,7 @@ ANTRAGSGRUEN_STRINGS = {
         "agendaAddEntry": "Add entry",
         "agendaDelEntryConfirm": "Delete this agenda item and all sub-items?",
         "removeAdminConfirm": "Do you really want to remove admin right from this user?",
+        "removeUserConfirm": "Do you really want to remove the user %NAME% from THIS consultation? Submitted motions etc. will remain in the system and need to be removed manually.",
         "emailMissingCode": "The text needs to contain the code %ACCOUNT%.",
         "emailMissingLink": "The text needs to contain the code %LINK%.",
         "emailMissingTo": "No e-mail-address was entered.",