+++ History.md
@@ -5,6 +5,7 @@
 - In the list of consultations, the newest consultation now appears at the top.
 - Clicking on changes in the public version of a merging draft now always opens a tooltip with the summary of the amendment, including the proposers of that amendment.
 - The custom status string is now also shown in the admin motion list.
+- When administering the list of users eligible to access a consultation, it is now possible to completely remove an user again.
 - Bugfix: the date of the last saved draft when merging motions was not set correctly on Safari.
 
 ## Version 4.1.0 (2018-11-17)